+++ app/controllers/internal_contacts_controller.rb
@@ -1,0 +1,10 @@
+class InternalContactsController < ApplicationController
+  before_action :find_project
+
+  def show
+  end
+
+  private def find_project
+    @project = Project.find(params[:project_id])
+  end
+end