+++ spec/features/users_can_view_project_information_spec.rb
@@ -55,23 +55,6 @@ RSpec.feature "Users can view project information" do
     end
   end
 
-  context "when the caseworker, team lead and regional delivery officer have been assigned" do
-    let(:project) { create(:conversion_project, :with_team_lead_and_regional_delivery_officer_assigned, caseworker: user) }
-
-    scenario "they can view the users names and email addresses assigned to the project" do
-      within("#projectDetails") do
-        expect(page).to have_content(user.full_name)
-        expect(page).to have_link("Email caseworker", href: "mailto:#{user.email}")
-
-        expect(page).to have_content(project.team_leader.full_name)
-        expect(page).to have_link("Email team leader", href: "mailto:#{project.team_leader.email}")
-
-        expect(page).to have_content(project.regional_delivery_officer.full_name)
-        expect(page).to have_link("Email regional delivery officer", href: "mailto:#{project.regional_delivery_officer.email}")
-      end
-    end
-  end
-
   context "when there are conditions from the advisory board" do
     let(:project) { create(:conversion_project, :with_conditions, caseworker: user) }