+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - the voluntary conversion, stakeholder kick off task had a duplicated action.
   This has now been corrected. Users may want to verify they have completed the
   action in the task.
+- renamed the "Contacts" tab to "External contacts"
+- moved the Caseworker, Team Leader and Regional Delivery Officer details to a
+  new "Internal contacts" tab.
 
 ## [Release 14][release-14]